Fixed the error in "Main" with our checkTime-method due to it only using one measurement "double time", whereas the method in "InputOutputProcess" needed a String name, LocalDate date, double time — so I changed it and now it works.

+++ src/Controller/InputOutputProcess.java
@@ -35,7 +35,7 @@ public class InputOutputProcess {
         calendar.checkDate(name, date);
     }
 
-    public void checkTime(String name, LocalDate date, int time) {
+    public void checkTime(String name, LocalDate date, double time) {
         calendar.checkTime(name, date, time);
     }
